refactor(front): move platform/auth DTO types out of lib/api into types/api
Tenth batch of the lib/api DTO migration: internal platform/auth endpoints.
Transport DTO types and zod schemas move to the mirrored path under types/api/*,
leaving business logic and domain types in place.

- 3 pure-contract files moved wholesale: groups, keys, auth_context.
- 3 files trimmed: oauth (moved credentials/bigquery/setup DTOs + schemas; kept
  OAuthError and all provider strategies), invitation (moved 8 DTO types/schemas;
  kept MembershipInvitationType and functions), workos/organization (moved only
  GetWorkspaceDomainsResponseBody).
- 24 consumers updated; precise per-statement rewrites preserve imports of stayed
  symbols (OAuthError, provider strategies, invitation/workos functions) on
  @app/lib; one mixed statement split.

Note: lib/api/credentials.ts (GetSlackClientIdResponseBody) deferred — its mirror
target types/api/credentials.ts is already occupied by an unrelated module;
needs a placement decision.
