+++ front-api/routes/invitations.ts
@@ -1,9 +1,9 @@
-import type { GetPendingInvitationsLookupResponseBody } from "@app/lib/api/invitation";
 import { fetchInvitationsFromOtherRegion } from "@app/lib/api/regions/lookup";
 import { getUserFromSession } from "@app/lib/iam/session";
 import { MembershipInvitationResource } from "@app/lib/resources/membership_invitation_resource";
 import { getMembershipInvitationToken } from "@app/lib/utils/invitation_token";
 import logger from "@app/logger/logger";
+import type { GetPendingInvitationsLookupResponseBody } from "@app/types/api/invitation";
 import type { PendingInvitationOption } from "@app/types/membership_invitation";
 import { sessionApp } from "@front-api/middlewares/ctx";
 import type { HandlerResult } from "@front-api/middlewares/utils";
