+++ front-api/routes/w/[wId]/keys/index.ts
@@ -3,14 +3,14 @@ import {
   emitAuditLogEvent,
   getAuditLogContext,
 } from "@app/lib/api/audit/workos_audit";
-import type {
-  GetKeysResponseBody,
-  PostKeysResponseBody,
-} from "@app/lib/api/keys";
 import { GroupResource } from "@app/lib/resources/group_resource";
 import { KeyResource } from "@app/lib/resources/key_resource";
 import { rateLimiter } from "@app/lib/utils/rate_limiter";
 import logger from "@app/logger/logger";
+import type {
+  GetKeysResponseBody,
+  PostKeysResponseBody,
+} from "@app/types/api/keys";
 import { workspaceApp } from "@front-api/middlewares/ctx";
 import { ensureIsAdmin } from "@front-api/middlewares/ensure_role";
 import { apiError, type HandlerResult } from "@front-api/middlewares/utils";
