+++ front/components/data_source/CreateOrUpdateConnectionBigQueryModal.tsx
@@ -1,11 +1,11 @@
 // Okay to use public API types because it's front/connectors communication.
 
 import { useTheme } from "@app/components/sparkle/ThemeContext";
-import type { PostCredentialsBody } from "@app/lib/api/oauth";
 import type { ConnectorProviderConfiguration } from "@app/lib/connector_providers";
 import { CONNECTOR_UI_CONFIGURATIONS } from "@app/lib/connector_providers_ui";
 import { clientFetch } from "@app/lib/egress/client";
 import { useBigQueryLocations } from "@app/lib/swr/bigquery";
+import type { PostCredentialsBody } from "@app/types/api/oauth";
 import type {
   ConnectorProvider,
   ConnectorType,
