+++ front/lib/api/oauth.ts
@@ -45,66 +45,12 @@ import type {
   OAuthProvider,
   OAuthUseCase,
 } from "@app/types/oauth/lib";
-import {
-  BigQueryCredentialsWithLocationSchema,
-  CheckBigQueryCredentialsSchema,
-  NotionCredentialsSchema,
-  SalesforceCredentialsSchema,
-  SnowflakeCredentialsSchema,
-} from "@app/types/oauth/lib";
 import type { OAuthAPIError } from "@app/types/oauth/oauth_api";
 import { OAuthAPI } from "@app/types/oauth/oauth_api";
 import type { Result } from "@app/types/shared/result";
 import { Err, Ok } from "@app/types/shared/result";
 import { isString } from "@app/types/shared/utils/general";
 import type { ParsedUrlQuery } from "querystring";
-import { z } from "zod";
-
-const PostSnowflakeCredentialsBodySchema = z.object({
-  provider: z.literal("snowflake"),
-  credentials: SnowflakeCredentialsSchema,
-});
-
-const PostBigQueryCredentialsBodySchema = z.object({
-  provider: z.literal("bigquery"),
-  credentials: BigQueryCredentialsWithLocationSchema,
-});
-
-const PostSalesforceCredentialsBodySchema = z.object({
-  provider: z.literal("salesforce"),
-  credentials: SalesforceCredentialsSchema,
-});
-
-const PostNotionCredentialsBodySchema = z.object({
-  provider: z.literal("notion"),
-  credentials: NotionCredentialsSchema,
-});
-
-export const PostCredentialsBodySchema = z.union([
-  PostSnowflakeCredentialsBodySchema,
-  PostBigQueryCredentialsBodySchema,
-  PostSalesforceCredentialsBodySchema,
-  PostNotionCredentialsBodySchema,
-]);
-
-export type PostCredentialsBody = z.infer<typeof PostCredentialsBodySchema>;
-export type PostCredentialsResponseBody = {
-  credentials: {
-    id: string;
-  };
-};
-
-export const PostCheckBigQueryRegionsRequestBodySchema = z.object({
-  credentials: CheckBigQueryCredentialsSchema,
-});
-
-export type PostCheckBigQueryLocationsResponseBody = {
-  locations: Record<string, string[]>;
-};
-
-export interface GetOAuthSetupResponseBody {
-  redirectUrl: string;
-}
 
 export type OAuthError = {
   code:
