+++ front/lib/api/workos/organization.ts
@@ -22,11 +22,6 @@ import {
 import assert from "assert";
 import uniqueId from "lodash/uniqueId";
 
-export interface GetWorkspaceDomainsResponseBody {
-  addDomainLink?: string;
-  domains: Organization["domains"];
-}
-
 export async function getOrCreateWorkOSOrganization(
   workspace: LightWorkspaceType,
   { domain }: { domain?: string } = {}
