+++ front/lib/swr/bigquery.ts
@@ -1,5 +1,5 @@
-import type { PostCheckBigQueryLocationsResponseBody } from "@app/lib/api/oauth";
 import { useFetcher, useSWRWithDefaults } from "@app/lib/swr/swr";
+import type { PostCheckBigQueryLocationsResponseBody } from "@app/types/api/oauth";
 import type { CheckBigQueryCredentials } from "@app/types/oauth/lib";
 import type { LightWorkspaceType } from "@app/types/user";
 import { useMemo } from "react";
