+++ front/lib/swr/workspaces.ts
@@ -27,10 +27,6 @@ import type {
   GetWorkspaceToolUsageResponse,
 } from "@app/lib/api/assistant/observability/tool_usage";
 import type { GetUserCreditsResponse } from "@app/lib/api/assistant/observability/user_credits";
-import type {
-  GetNoWorkspaceAuthContextResponseType,
-  GetWorkspaceAuthContextResponseType,
-} from "@app/lib/api/auth_context";
 import type { GetBillingInfoResponseBody } from "@app/lib/api/billing/info";
 import type { GetBillingInvoicesResponseBody } from "@app/lib/api/billing/invoices";
 import type {
@@ -39,7 +35,6 @@ import type {
 } from "@app/lib/api/checkout/business_activation";
 import type { GetPreparePaymentResponseBody } from "@app/lib/api/checkout/prepare_payment";
 import type { GetMetronomeContractResponseBody } from "@app/lib/api/credits/metronome_contract";
-import type { GetPendingInvitationsLookupResponseBody } from "@app/lib/api/invitation";
 import type {
   GetCheckoutStatusResponseBody,
   GetSubscriptionsResponseBody,
@@ -73,6 +68,11 @@ import type {
 } from "@app/lib/resources/subscription_resource";
 import type { GetJoinResponseBody } from "@app/lib/signup";
 import { emptyArray, useFetcher, useSWRWithDefaults } from "@app/lib/swr/swr";
+import type {
+  GetNoWorkspaceAuthContextResponseType,
+  GetWorkspaceAuthContextResponseType,
+} from "@app/types/api/auth_context";
+import type { GetPendingInvitationsLookupResponseBody } from "@app/types/api/invitation";
 import type { APIErrorResponse, RegionRedirectError } from "@app/types/error";
 import type { BillingPeriod } from "@app/types/plan";
 import { safeParseJSON } from "@app/types/shared/utils/json_utils";
