+++ front/types/api/workos/organization.ts
@@ -1,0 +1,6 @@
+import type { Organization } from "@workos-inc/node";
+
+export interface GetWorkspaceDomainsResponseBody {
+  addDomainLink?: string;
+  domains: Organization["domains"];
+}
